WIP: added relationship renders and change contents rendering
Lots of change in this commit. this should actually be devided into
multiple commits, but im too lazy to split them.
The idea was to list contents per package so files are bundled by pkg id
instead now each file will have its own entity/id and can be listed individually.
This will make it easier to apply filters to files/contents but the output will
be much more verbose.